+++ streamplugins/01_somafm.php
@@ -31,12 +31,8 @@ class somafmplugin {
         print '<div class="containerbox padright indent ninesix bumpad brick_wide">';
         print '<a href="http://somafm.com" target="_blank">'.get_int_text("label_soma_beg").'</a>';
         print '</div>';
-        $d = new url_downloader(array('url' => "http://api.somafm.com/channels.xml"));
-        if ($d->get_data_to_string()) {
-            $this->doAllStations($d->get_data());
-        } else {
-            print 'There was an error getting the channels from Soma FM - status code '.$d->get_status();
-        }
+        $stations = getCacheData("http://api.somafm.com/channels.xml", 'somafm', true, true);
+        $this->doAllStations($stations);
     }
 
     // -- Private Functions -- //
@@ -78,16 +74,20 @@ class somafmplugin {
 
     private function doAllStations($content) {
         logger::trace("SOMAFM", "Loaded Soma FM channels list");
-        $x = simplexml_load_string($content);
-        $count = 0;
-        foreach ($x->channel as $channel) {
-            $this->doChannel($count, $channel);
-            $count++;
+        try {
+            $x = simplexml_load_string($content);
+            $count = 0;
+            foreach ($x->channel as $channel) {
+                $this->doChannel($count, $channel);
+                $count++;
+            }
+        } catch (Exception $e) {
+            print 'There was an error getting the channels from Soma FM';
         }
     }
 
     private function doChannel($count, $channel) {
-        logger::log("SOMAFM", "Channel :",$channel->title);
+        logger::log("SOMAFM", "Channel :", (string) $channel->title);
         if ($channel->highestpls) {
             $pls = (string) $channel->highestpls;
         } else {